refactor(table): make tableRole type available in all extensions that use it
