+++ packages/extension-table/src/cell/table-cell.ts
@@ -1,3 +1,5 @@
+import '../types.js'
+
 import { mergeAttributes, Node } from '@tiptap/core'
 
 export interface TableCellOptions {
