+++ packages/extension-table/src/header/table-header.ts
@@ -1,3 +1,5 @@
+import '../types.js'
+
 import { mergeAttributes, Node } from '@tiptap/core'
 
 export interface TableHeaderOptions {
