+++ packages/extension-table/src/row/table-row.ts
@@ -1,3 +1,5 @@
+import '../types.js'
+
 import { mergeAttributes, Node } from '@tiptap/core'
 
 export interface TableRowOptions {
