+++ packages/extension-table/src/types.ts
@@ -1,0 +1,19 @@
+import { ParentConfig } from '@tiptap/core'
+
+declare module '@tiptap/core' {
+  interface NodeConfig<Options, Storage> {
+    /**
+     * A string or function to determine the role of the table.
+     * @default 'table'
+     * @example () => 'table'
+     */
+    tableRole?:
+      | string
+      | ((this: {
+          name: string
+          options: Options
+          storage: Storage
+          parent: ParentConfig<NodeConfig<Options>>['tableRole']
+        }) => string)
+  }
+}
